Modified wsc impl. to reconnect fast on startup
websocketclientimplementation now calls onDisconnected() SLOT right
after startup. If this was not done, it takes a long time until
SLOT is triggered from QWebSocket. Now the client tries to reconnect
immediately, if the server is not reachable on startup. Before this
change, it could take almost a minute until a later starting server
was connected.