Modified ColorRecognitionUnit.qml's state update
Before the state of ColorRecognitionUnit, after a color has been
recognized was not updated correctly. This has been changed.
Additionally the state is reset after one second, to have a proper
animation in the GUI.
As a consequence, the files websocketclient.h/.cpp have been
modified to have a proper SIGNAL for this purpose.